# docs/openapi/accounts.yaml
openapi: 3.0.3
info:
  title: Bank Account API
  version: 1.0.0

paths:
  /v1/accounts:
    post:
      summary: Create a new bank account
      operationId: createAccount
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - insufficient scope

  /v1/accounts/{id}:
    get:
      summary: Get account details
      operationId: getAccount
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CreateAccountRequest:
      type: object
      required:
        - owner
      properties:
        owner:
          type: string
    AccountResponse:
      type: object
      properties:
        id:
          type: string
        owner:
          type: string
        balance:
          type: integer
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        message:
          type: string